Add underlying Ns to a tabyl displaying percentages.

This function adds back the underlying Ns to a tabyl whose percentages were calculated using adorn_percentages(), to display the Ns and percentages together. You can also call it on a non-tabyl data.frame with tabyl-like format to which you wish to append Ns.

adorn_ns(dat, position = "rear", ns = attr(dat, "core"))

Arguments

dat
a data.frame of class tabyl that has had adorn_percentages and/or adorn_pct_formatting called on it.

position
should the N go in the front, or in the rear, of the percentage?

ns
the Ns to append. The default is the "core" attribute of the input tabyl dat, where the original Ns of a two-way tabyl are stored. However, if you need to modify the numbers, e.g., to format 4000 as 4,000 or 4k, you can do that separately and supply the formatted result here.

Value

a data.frame with Ns appended

Examples

mtcars %>%
  tabyl(am, cyl) %>%
  adorn_percentages("col") %>%
  adorn_pct_formatting() %>%
  adorn_ns(position = "front")

#>  am         4         6          8
#>   0 3 (27.3%) 4 (57.1%) 12 (85.7%)
#>   1 8 (72.7%) 3 (42.9%)  2 (14.3%)