fisher.test {janitor}	R Documentation

Apply stats::fisher.test to a two-way tabyl

Description

This generic function overrides stats::fisher.test. If the passed table is a two-way tabyl, it runs it through janitor::fisher.test.tabyl, otherwise it just calls stats::fisher.test.

Usage

fisher.test(x, ...)

## Default S3 method:
fisher.test(x, y = NULL, ...)

## S3 method for class 'tabyl'
fisher.test(x, ...)

Arguments

x
a two-way tabyl, a numeric vector or a factor

...
other parameters passed to stats::fisher.test

y
if x is a vector, must be another vector or factor of the same length

Value

The result is the same as the one of stats::fisher.test.

Examples

tab <- tabyl(mtcars, gear, cyl)
fisher.test(tab)

[Package janitor version 2.2.1 Index]