janitor: Simple Tools for Examining and Cleaning Dirty Data

The main janitor functions can: perfectly format data.frame column names; provide quick counts of variable combinations (i.e., frequency tables and crosstabs); and explore duplicate records. Other janitor functions nicely format the tabulation results. These tabulate-and-report functions approximate popular features of SPSS and Microsoft Excel. This package follows the principles of the "tidyverse" and works well with the pipe function

Package context

Advanced users can do most things covered here, but they can do it faster with janitor and save their thinking for more fun tasks.

See also

Author

Maintainer: Sam Firke samuel.firke@gmail.com

Other contributors:

Bill Denney wdenney@humanpredictions.com [contributor]

Chris Haid chrishaid@gmail.com [contributor]

Ryan Knight ryangknight@gmail.com [contributor]

Malte Grosser malte.grosser@gmail.com [contributor]

Jonathan Zadra jonathan.zadra@sorensonimpact.com [contributor]

Olivier Roy [contributor]