janitor_deprecated {janitor}	R Documentation

Deprecated Functions in Package janitor

Description

These functions are provided for compatibility with older versions of janitor only, and may be defunct as soon as the next release.

Details

use_first_valid_of

convert_to_NA

[Package janitor version 1.0.0 Index]