mu_to_u {janitor}	R Documentation

Constant to help map from mu to u

Description

This is a character vector with names of all known Unicode code points that look like the Greek mu or the micro symbol and values of "u". This is intended to simplify mapping from mu or micro in Unicode to the character "u" with clean_names() and make_clean_names().

Usage

mu_to_u

Format

An object of class character of length 10.

Details

See the help in clean_names() for how to use this.

See Also

Other Set names: clean_names(), find_header(), row_to_names()

[Package janitor version 2.2.1 Index]