Convert a data.frame of counts to percentages.

This function excludes the first column of the input data.frame, assuming that it contains a descriptive variable. If the input data.frame is not a tabyl, it will convert to one in order to preserve the underlying values in the core attribute.

adorn_percentages(dat, denominator = "row", na.rm = TRUE)

Arguments

dat
a tabyl or other data.frame with a tabyl-like layout. If given a list of data.frames, this function will apply itself to each data.frame in the list (designed for 3-way tabyl lists).

denominator
the direction to use for calculating percentages. One of "row", "col", or "all".

na.rm
should missing values (including NaN) be omitted from the calculations?

Value

Returns a data.frame of percentages, expressed as numeric values between 0 and 1.

Examples

mtcars %>%
  tabyl(am, cyl) %>%
  adorn_percentages("col")

#>  am         4         6         8
#>   0 0.2727273 0.5714286 0.8571429
#>   1 0.7272727 0.4285714 0.1428571

# calculates correctly even with totals column and/or row:
mtcars %>%
  tabyl(am, cyl) %>%
  adorn_totals("row") %>%
  adorn_percentages()

#>     am         4         6         8
#>      0 0.1578947 0.2105263 0.6315789
#>      1 0.6153846 0.2307692 0.1538462
#>  Total 0.3437500 0.2187500 0.4375000