Add tabyl attributes to a data.frame

A tabyl is a data.frame containing counts of a variable or co-occurrences of two variables (a.k.a., a contingency table or crosstab). This specialized kind of data.frame has attributes that enable adorn_ functions to be called for precise formatting and presentation of results. E.g., display results as a mix of percentages, Ns, add totals rows or columns, rounding options, in the style of Microsoft Excel PivotTable.

A tabyl can be the result of a call to janitor::tabyl(), in which case these attributes are added automatically. This function adds tabyl class attributes to a data.frame that isn't the result of a call to tabyl but meets the requirements of a two-way tabyl: 1) First column contains values of variable 1 2) Column names 2:n are the values of variable 2 3) Numeric values in columns 2:n are counts of the co-occurrences of the two variables.*

= this is the ideal form of a tabyl, but janitor's adorn_ functions tolerate and ignore non-numeric columns in positions 2:n.

For instance, the result of dplyr::count() followed by tidyr::pivot_wider() can be treated as a tabyl.

The result of calling tabyl() on a single variable is a special class of one-way tabyl; this function only pertains to the two-way tabyl.

Usage

as_tabyl(dat, axes = 2, row_var_name = NULL, col_var_name = NULL)

Arguments

dat

a data.frame with variable values in the first column and numeric values in all other columns.

axes

is this a two_way tabyl or a one_way tabyl? If this function is being called by a user, this should probably be "2". One-way tabyls are created by tabyl but are a special case.

row_var_name

(optional) the name of the variable in the row dimension; used by adorn_title().

col_var_name

(optional) the name of the variable in the column dimension; used by adorn_title().

Value

Returns the same data.frame, but with the additional class of "tabyl" and the attribute "core".

Examples

as_tabyl(mtcars)
#>   mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#>  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#>  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#>  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#>  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#>  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#>  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#>  14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#>  24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#>  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#>  19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#>  17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#>  16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#>  17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#>  15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#>  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#>  10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#>  14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#>  32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#>  30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#>  33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#>  21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#>  15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#>  15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#>  13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#>  19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#>  27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#>  26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#>  30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#>  15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#>  19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#>  15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#>  21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2