Get rows of a data.frame with identical values for the specified variables.

For hunting duplicate records during data cleaning. Specify the data.frame and the variable combination to search for duplicates and get back the duplicated rows.

get_dupes(dat, ...)

Arguments

dat
The input data.frame.

...
Unquoted variable names to search for duplicates. This takes a tidyselect specification.

Value

Returns a data.frame with the full records where the specified variables have duplicated values, as well as a variable dupe_count showing the number of rows sharing that combination of duplicated values. If the input data.frame was of class tbl_df, the output is as well.

Examples

get_dupes(mtcars, mpg, hp)

#> # A tibble: 2 x 12
#>     mpg    hp dupe_count   cyl  disp  drat    wt  qsec    vs    am  gear  carb
#>   <dbl> <dbl>      <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1    21   110          2     6   160   3.9  2.62  16.5     0     1     4     4
#> 2    21   110          2     6   160   3.9  2.88  17.0     0     1     4     4

# or called with the magrittr pipe %>% :
mtcars %>% get_dupes(wt)

#> # A tibble: 5 x 12
#>      wt dupe_count   mpg   cyl  disp    hp  drat  qsec    vs    am  gear  carb
#>   <dbl>      <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  3.44          3  18.7     8  360    175  3.15  17.0     0     0     3     2
#> 2  3.44          3  19.2     6  168.   123  3.92  18.3     1     0     4     4
#> 3  3.44          3  17.8     6  168.   123  3.92  18.9     1     0     4     4
#> 4  3.57          2  14.3     8  360    245  3.21  15.8     0     0     3     4
#> 5  3.57          2  15       8  301    335  3.54  14.6     0     1     5     8

# You can use tidyselect helpers to specify variables:
mtcars %>% get_dupes(-c(wt, qsec))

#> # A tibble: 2 x 12
#>     mpg   cyl  disp    hp  drat    vs    am  gear  carb dupe_count    wt  qsec
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>      <int> <dbl> <dbl>
#> 1    21     6   160   110   3.9     0     1     4     4          2  2.62  16.5
#> 2    21     6   160   110   3.9     0     1     4     4          2  2.88  17.0

mtcars %>% get_dupes(starts_with("cy"))

#> # A tibble: 32 x 12
#>      cyl dupe_count   mpg  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl>      <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     4         11  22.8 108      93  3.85  2.32  18.6     1     1     4     1
#>  2     4         11  24.4 147.     62  3.69  3.19  20       1     0     4     2
#>  3     4         11  22.8 141.     95  3.92  3.15  22.9     1     0     4     2
#>  4     4         11  32.4  78.7    66  4.08  2.2   19.5     1     1     4     1
#>  5     4         11  30.4  75.7    52  4.93  1.62  18.5     1     1     4     2
#>  6     4         11  33.9  71.1    65  4.22  1.84  19.9     1     1     4     1
#>  7     4         11  21.5 120.     97  3.7   2.46  20.0     1     0     3     1
#>  8     4         11  27.3  79      66  4.08  1.94  18.9     1     1     4     1
#>  9     4         11  26   120.     91  4.43  2.14  16.7     0     1     5     2
#> 10     4         11  30.4  95.1   113  3.77  1.51  16.9     1     1     5     2
#> # ... with 22 more rows