janitor

janitor has simple little tools for examining and cleaning dirty data.

Main functions

The main janitor functions can: perfectly format ugly data.frame column names; isolate duplicate records for further study; and provide quick one- and two-variable tabulations (i.e., frequency tables and crosstabs) that improve on the base R function table().

Other functions in the package can format for reporting the results of these tabulations. These tabulate-and-report functions approximate popular features of SPSS and Microsoft Excel.

Package context

This package follows the principles of the "tidyverse" and in particular works well with the %>% pipe function.

janitor was built with beginning-to-intermediate R users in mind and is optimized for user-friendliness. Advanced users can already do everything covered here, but they can do it faster with janitor and save their thinking for more fun tasks.