top_levels {janitor}	R Documentation

Generate a frequency table of a factor grouped into top-n, bottom-n, and all other levels.

Description

Get a frequency table of a factor variable, grouped into categories by level.

Usage


top_levels(input_vec, n = 2, show_na = FALSE)

Arguments

input_vec
the factor variable to tabulate.

n
number of levels to include in top and bottom groups

show_na
should cases where the variable is NA be shown?

Value

Returns a data.frame (actually a tbl_df) with the frequencies of the grouped, tabulated variable. Includes counts and percentages, and valid percentages (calculated omitting NA values, if present in the vector and show_na = TRUE.)

Examples


top_levels(as.factor(mtcars$hp), 2)

[Package janitor version 1.0.0 Index]